Move 'app' module to 'com.adobe.marketing.mobile.edge.testapp.java' namespace
