+++ code/app/src/main/AndroidManifest.xml
@@ -18,14 +18,14 @@
     <application
         android:allowBackup="true"
         android:label="@string/app_name"
-        android:name="TestApplication"
+        android:name=".TestApplication"
         tools:replace="android:icon,android:roundIcon"
         android:icon="@drawable/aep_logo"
         android:roundIcon="@drawable/aep_logo"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:networkSecurityConfig="@xml/network_security_config">
-        <activity android:name="com.adobe.marketing.tester.MainActivity"
+        <activity android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -52,7 +52,7 @@
             android:theme="@style/AppTheme">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.adobe.marketing.tester.AssuranceActivity" />
+                android:value=".AssuranceActivity" />
         </activity>
     </application>
